fix: some bug fix

+++ setup.sh
@@ -16,19 +16,19 @@ RE_INSTALL=false
 
 # 询问是否安装（如果未启用 -a）
 ask_install() {
-    sleep 1
+    sleep 3
     clear
     local package="$1"
 
     # 如果已安装，直接跳过
-    if command -v "$package" &> /dev/null && ! RE_INSTALL ; then
-        echo -e "\n\033[1;32m✅ $package 已安装，跳过...\033[0m"
+    if command -v "$package" &> /dev/null && ! $RE_INSTALL ; then
+        echo -e "\033[1;32m✅ $package 已安装，跳过...\033[0m"
         return 1
     fi
 
     # 如果有 -a 则不询问
     if $AUTO_INSTALL; then
-        echo -e "\n\033[1;36m🔹 正在安装 $package...\033[0m"
+        echo -e "\033[1;36m🔹 正在安装 $package...\033[0m"
         return 0
     fi
 
@@ -48,14 +48,15 @@ chmod +x ./install_tools.sh && ./install_tools.sh
 ask_install "stow" && sudo apt update && sudo apt install -y stow
 
 # shell
+sleep 3
 clear
-if command -v zsh > /dev/null; then
-    echo -e "\n\033[1;32m✅ zsh 已安装，跳过...\033[0m"
+if command -v zsh > /dev/null && ! $RE_INSTALL ; then
+    echo -e "\033[1;32m✅ zsh 已安装，跳过...\033[0m"
     rm -f ~/.zshrc
     rm -f ~/.bashrc
     stow_module "shell"
 else
-    echo -e "\n\033[1;36m🔹 正在安装 shell, 请选择要进行的操作\033[0m"
+    echo -e "\033[1;36m🔹 正在安装 shell, 请选择要进行的操作\033[0m"
     echo "[1] 安装 zsh 并设为默认终端"
     echo "[2] 安装 zsh 但依然使用 bash"
     echo "[3] 不安装 zsh"
@@ -65,9 +66,10 @@ else
     case $shell_choice in
         1)
             sudo apt install -y zsh
+	    rm -rf "$HOME/.oh-my-zsh"
             sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
 
-            ZSH_CUSTOM="$HOME/.oh-my-zsh/custom/plugins"
+            ZSH_CUSTOM="$HOME/.oh-my-zsh/plugins"
             git clone https://github.com/zsh-users/zsh-syntax-highlighting.git "$ZSH_CUSTOM/zsh-syntax-highlighting"
             git clone https://github.com/zsh-users/zsh-autosuggestions.git "$ZSH_CUSTOM/zsh-autosuggestions"
             git clone https://github.com/jeffreytse/zsh-vi-mode.git "$ZSH_CUSTOM/zsh-vi-mode"
@@ -79,6 +81,7 @@ else
             ;;
         2)
             sudo apt install -y zsh
+	    rm -rf "$HOME/.oh-my-zsh"
             sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
 
             ZSH_CUSTOM="$HOME/.oh-my-zsh/custom/plugins"
@@ -126,6 +129,6 @@ if ask_install "tmux"; then
     git clone https://github.com/tmux-plugins/tpm "$HOME/.tmux/plugins/tpm"
 fi
 
-sleep 1
+sleep 3
 clear
-echo -e "\n\033[1;32m✅ Dotfiles 配置完成！\033[0m"
+echo -e "\033[1;32m✅ Dotfiles 配置完成！\033[0m"
